TravelAlly-Backend/requirements.txt
Django>=3.0.4,<3.1.0
djangorestframework>=3.11.0,<3.12.0
psycopg2>=2.8.4,<2.9.0
Pillow>=5.3.0,<5.4.0

flake8>=3.7.9,<3.8.0
